[리팩토링] 테스트 코드 설정 변경 및 패키지 이동
