--- src/main/java/com/projectteam/coop/domain/AttachmentStatus.java
@@ -1,4 +1,0 @@
-package com.projectteam.coop.domain;
-
-public enum AttachmentStatus {
-}
