--- src/main/java/com/projectteam/coop/domain/MemberType.java
@@ -1,5 +1,0 @@
-package com.projectteam.coop.domain;
-
-public enum MemberType {
-    ADMIN, COMMON
-}
