--- src/main/java/com/projectteam/coop/domain/PostStatus.java
@@ -1,5 +1,0 @@
-package com.projectteam.coop.domain;
-
-public enum PostStatus {
-    Y, N
-}
