+++ src/main/java/com/projectteam/coop/domain/board/enums/AttachmentStatus.java
@@ -1,0 +1,4 @@
+package com.projectteam.coop.domain.board.enums;
+
+public enum AttachmentStatus {
+}
