+++ src/main/java/com/projectteam/coop/domain/board/enums/BoardStatus.java
@@ -1,0 +1,5 @@
+package com.projectteam.coop.domain.board.enums;
+
+public enum BoardStatus {
+    Y, N
+}
