+++ src/main/java/com/projectteam/coop/domain/board/enums/BoardType.java
@@ -1,0 +1,5 @@
+package com.projectteam.coop.domain.board.enums;
+
+public enum BoardType {
+    TEXT, IMG
+}
