+++ src/main/java/com/projectteam/coop/domain/member/enums/MemberStatus.java
@@ -1,0 +1,5 @@
+package com.projectteam.coop.domain.member.enums;
+
+public enum MemberStatus {
+    ACTIVE
+}
