+++ src/main/java/com/projectteam/coop/domain/member/enums/MemberType.java
@@ -1,0 +1,5 @@
+package com.projectteam.coop.domain.member.enums;
+
+public enum MemberType {
+    ADMIN, COMMON
+}
