+++ src/main/java/com/projectteam/coop/domain/post/enums/PostStatus.java
@@ -1,0 +1,5 @@
+package com.projectteam.coop.domain.post.enums;
+
+public enum PostStatus {
+    Y, N
+}
