+++ src/main/java/com/projectteam/coop/domain/product/enums/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.projectteam.coop.domain.product.enums;
+
+public enum ProductStatus {
+    Y, N
+}
