+++ src/test/java/com/projectteam/coop/service/login/LoginServiceTest.java
@@ -1,23 +1,19 @@
 package com.projectteam.coop.service.login;
 
-import com.projectteam.coop.domain.member.model.entity.Member;
 import com.projectteam.coop.domain.log.LoginLog;
 import com.projectteam.coop.domain.login.service.LoginService;
-import com.projectteam.coop.repository.log.LoginRepository;
+import com.projectteam.coop.domain.member.model.entity.Member;
 import com.projectteam.coop.domain.member.service.MemberService;
+import com.projectteam.coop.repository.log.LoginRepository;
 import com.projectteam.coop.util.SecurityUtil;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
